Adjusted v3 to reflect that GHG is a positive indicator (higher=better) and to change the folder structure.
Adjusted V4 to change the folder structure and allow to run the code for HICs.

There is a mistake with Lebanon in the data downloaded from WDI - this will need to be fixed manually in the code (as of now when V3 is run the error is introduced each time).